Add files via upload
This tables represent all the tables included in my PhD thesis. The tables are named according to the chapter in which they appear. The S before the numerical means supplementary while the number immediately after S represents the chapter. For example, Table S2.1 is found in chapter 2 and is the first table.